chore(deps): Revert - Upgrade jol-core to latest
This reverts the jol-core update that broke old jdk compatibility

+++ stats/src/main/java/com/facebook/airlift/stats/cardinality/HllInstance.java
@@ -26,7 +26,7 @@ interface HllInstance
 
     int getIndexBitLength();
 
-    long estimatedInMemorySize();
+    int estimatedInMemorySize();
 
     int estimatedSerializedSize();
